final update

+++ app/src/main/java/com/app/shopinkarts/response/DealOfDay.kt
@@ -17,7 +17,7 @@ data class DealOfDay(
     val tax: Int,
     val taxType: Int,
     val variantsArr: List<VariantsArr>,
-    val avgRating: Int,
+    val avgRating: Double,
     val deliveryInstructions: List<String>,
     val dispatchDetails: List<String>,
     val isFreeDelivery: Int,
