+++ app/src/main/java/com/app/shopinkarts/response/Product.kt
@@ -3,7 +3,7 @@ package com.app.shopinkarts.response
 data class Product(
     val _id: String,
     val attributes: List<Attribute>,
-    val avgRating: Int,
+    val avgRating: Float,
     val categoryId: String,
     val hsnCode: String,
     val creationTimeStamp: String,
