+++ app/src/main/java/com/app/shopinkarts/response/YouMayLike.kt
@@ -4,7 +4,7 @@ package com.app.shopinkarts.response
 data class YouMayLike(
     val _id: String,
     val attributes: List<Attribute>,
-    val avgRating: Int,
+    val avgRating: Double,
     val categoryId: String,
     val creationTimeStamp: String,
     val deliveryInstructions: List<String>,
